+++ lib/rails_tracepoint_stack/config.rb
@@ -1,0 +1,9 @@
+module RailsTracepointStack
+  class Config
+    attr_accessor :ignore_patterns
+
+    def initialize
+      @ignore_patterns = []
+    end
+  end
+end
